Don't generate gpg if it is already there.

+++ scripts/submit.sh
@@ -3,8 +3,6 @@
 # Script to prepare the output files for submission. By pairs, it tars them, encrypts them and ftps them
 # NOTE: this script isn't very robust, use with care and insight.
 #
-# Currently, the script only submits ASMT* files;
-#
 # The gpg environment must be set up to allow encrypt/sign by a known user for the proper recipient.
 # Make sure that is so and modify the gpg command as necessary. For example, if there is a single private
 # key configured in the config/gpg folder and the recipient is the usual production value:
@@ -20,18 +18,21 @@ pushd "$SCRIPT_DIR/../out"
 
 for csv in ASMT*.csv; do
     base="${csv%.*}"
-    if [ ! -f "$base.json" ]; then
-        echo "skipping unpaired file $csv"
-    else
-        tarfile="$base.tar.gz"
-        gpgfile="$tarfile.gpg"
-        tar czf "$tarfile" "$base.csv" "$base.json"
-        gpg --batch --no-tty --yes -q --passphrase ca_user -u ca_user -r sbacdw -es -o "$gpgfile" "$tarfile"
-        rm "$tarfile"
-        echo "submitting $gpgfile"
-        "$SCRIPT_DIR/submit.exp" "$gpgfile"
-#        sleep 2
+    tarfile="$base.tar.gz"
+    gpgfile="$tarfile.gpg"
+    if [ ! -f "$gpgfile" ]; then
+        if [ ! -f "$base.json" ]; then
+            echo "skipping unpaired file $csv"
+            continue
+        else
+            tar czf "$tarfile" "$base.csv" "$base.json"
+            gpg --batch --no-tty --yes -q --passphrase ca_user -u ca_user -r sbacdw -es -o "$gpgfile" "$tarfile"
+            rm "$tarfile"
+        fi
     fi
+    echo "submitting $gpgfile"
+    "$SCRIPT_DIR/submit.exp" "$gpgfile"
+#    sleep 2
 done
 
 popd
